# src/main/java/org/javig/micro_neural/model/MundoRequest.java
package org.javig.micro_neural.model;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class MundoRequest {

    private int[][] data;
    private int turno;

}